actions: Split out zerver.actions.message_flags.
(cherry picked from commit eb4e9fe1e73067de3cfc3b95b061763815169ec6)
